Fix #24
The use of `{{- with ... }}` seems useless here, and it breaks everything else. I think(?) it doesn't cause an issue in the default configuration because `.Values.networkPolicy.ingress.grpc` is an empty array (and I guess that means it doesn't actually switch into that scope?).

Would be nice™ to add a test too (I made this PR from the GitHub web app).