feat: make Init Job annotations customizable
In some cases, there is a need for adding custom annotations to the Init job.
For example, when using ArgoCD, we need to instruct ArgoCD to start the job
still during the Sync phase and not as a PostSync hook. Otherwise, the job never
starts. And this is done by adding the following annotation to the job.
```
argocd.argoproj.io/hook: Sync
```

This is to continue the work started in https://github.com/cockroachdb/helm-charts/pull/246
to address https://github.com/cockroachdb/helm-charts/issues/149