test: fix flaky test due to timeout
The e2e test times out probably due to a lot of PVs getting provisioned
and running out of space.

This PR restricts the PV sizes across all tests to 1Gi (as opposed to
the default 100Gi). It also now uses the ubuntu-latest-4-core runner
(with higher storage) for the e2e tests.