Manage Transparent Huge Pages for CRDB
THP is unhelpful for databases, in general, because memory access tends
to be non-contiguous, and the use of THP can make performance worse when
it requires a bunch of page defragmentation to load and map a huge page.

We disable it by default, but allow the chart user to modify the policy.