Use the GODEBUG env var to disable Transparent Huge Pages (#463)
THP is unhelpful for databases, in general, because memory access tends to be non-contiguous, and the use of THP can make performance worse when it requires a bunch of page defragmentation to load and map a huge page.

We now disable it by default, using the GODEBUG environment variable.

Adds a generic facility for managing GODEBUG clauses to the Helm chart.

closes https://cockroachlabs.atlassian.net/browse/TREQ-888