Visus Sidecar config (#507)
* Visus Sidecar config

First-class support for Visus as a CRDB Sidecar in the Helm chart.

Visus will start before the DB, and be restarted if it fails. K8s won't
wait for visus to terminate, since the restartPolicy is set.

Also includes a separate Service def so the Prometheus scraper can
address each node directly.